save work on khanh_ver_@

+++ NCF/src/scripts/genericNeuralNet.py
@@ -47,8 +47,7 @@ def variable_with_weight_decay(name, shape, stddev, wd):
         name, 
         shape, 
         initializer=tf.truncated_normal_initializer(
-            stddev=stddev, 
-            dtype=dtype))
+            stddev=stddev))
  
     if wd is not None:
       weight_decay = tf.multiply(tf.nn.l2_loss(var), wd, name='weight_loss')
